Moved Auth part from web to nonzone-lib
Auth Provider and hooks for user, balance, profile are now exported by the library.

Removed redux from webapp as it doesn't seem needed for the moment
