+++ webapp/src/main_components/Create.js
@@ -10,7 +10,7 @@ import {
     getPublishPrice,
     getCurrency,
 } from 'nonzone-lib';
-import { useAuth, useUserWallet } from '../Auth';
+import { useAuth, useUserWallet } from 'nonzone-lib';
 import { Spinner } from '../components/spinner/Spinner';
 
 const MIN_DESCR_LENGTH = 150;
