+++ webapp/src/main_components/Nonzone.js
@@ -3,7 +3,7 @@ import { Interface, svg, Sliderr, Spinner, DialogWindow } from '../components';
 import './nonzone.css';
 import { useParams } from 'react-router-dom';
 import { useLoadStory, sendTip, getCurrency, Login } from 'nonzone-lib';
-import { useAuth, useUserWallet } from '../Auth';
+import { useAuth, useUserWallet } from 'nonzone-lib';
 
 const TIP_AMOUNT = 0.01;
 const MIN_FUNDS_FOR_TIP = TIP_AMOUNT + 0.00000001;
