+++ webapp/src/main_components/ProfileWidget.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useAuth } from '../Auth';
+import { useAuth } from 'nonzone-lib';
 import { Profile } from '../components';
 
 export const ProfileWidget = ({ onShowProfile, onSignIn }) => {
